requirements.txt
pandas
sentence-transformers
faiss-cpu
sqlalchemy
python-dotenv
tqdm
requests
openai
huggingface-hub
streamlit